Split up preprocessors
